Address clint warnings and other style issues.
* Add const.
* Fix conditions (move && from end to start of line).
* Use int32_t instead of long.
* Use //-style comments.

+++ src/os/input.h
@@ -11,9 +11,8 @@ bool input_ready(void);
 void input_start(void);
 void input_stop(void);
 uint32_t input_read(char *buf, uint32_t count);
-int os_inchar(char_u *, int, long, int);
+int os_inchar(char_u *, int, int32_t, int);
 bool os_char_avail(void);
 void os_breakcheck(void);
 
-#endif
-
+#endif  // NEOVIM_OS_INPUT_H
